requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-eventhub==5.11.2
azure-functions==1.13.3
azure-identity==1.12.0
azure-mgmt-cognitiveservices==13.4.0
azure-mgmt-core==1.3.2
azure-mgmt-resource==22.0.0
matplotlib==3.7.1
openai==0.27.2
pandas==1.5.3
plotly==5.13.1
pyodbc==4.0.35
redis==4.5.4
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
tenacity==8.2.2
tiktoken==0.3.0